+++ dicts/src/test/java/com/backstage/app/dict/service/postgres/PostgresDictDataServiceTest.java
@@ -53,6 +53,24 @@ public class PostgresDictDataServiceTest extends CommonDictDataServiceTest
 	}
 
 	@Test
+	protected void check_streamByFilter()
+	{
+		streamByFilter();
+	}
+
+	@Test
+	protected void check_checkSerialField()
+	{
+		checkSerialField();
+	}
+
+	@Test
+	protected void check_checkReserverWords()
+	{
+		checkReserverWords();
+	}
+
+	@Test
 	void getByFilterWithNullRefFieldCorrect()
 	{
 		getByFilterWithNullRefField();